Add access restrictions to timers and time records with cancancan gem and start to make correct filter

+++ app/controllers/timers_controller.rb
@@ -6,17 +6,14 @@ class TimersController < ApplicationController
     respond_to do |format|
       format.html # index.html.erb
       format.json do
-        records = TimeRecord.ordered.limit(12)
-        unless current_user.admin?
-          records = records.where(nakama: current_nakama)
-        end
+        records = TimeRecord.accessible_by(current_ability, :read).ordered.limit(12)
         timers = current_user.timers
 
         render json: {
           records: records,
           timers: timers.map { |timer| timer_json_data(timer) },
           isAdmin: current_user.admin?,
-          nakama: Nakama.all.collect,
+          nakama: Nakama.accessible_by(current_ability, :read).collect,
           nakamaId: current_user.nakama_id
         }
       end
@@ -24,43 +21,63 @@ class TimersController < ApplicationController
   end
 
   def create
-    timer = Timer.new user: current_user
-    timer.assign_attributes(timer_params)
-    timer.started_at = Time.now
-    timer.save
-    render json: { timer: timer_json_data(timer) }
+    if can? :create, Timer
+      timer = Timer.new user: current_user
+      timer.assign_attributes(timer_params)
+      timer.started_at = Time.now
+      timer.save
+      render json: { timer: timer_json_data(timer) }
+    else
+      render json: { error: 'access denied' }, status: :unprocessable_entity
+    end
   end
 
   def update
-    timer = current_user.timers.find params[:id]
-    if params[:act] == 'play'
-      if timer.started_at.nil?
-        timer.started_at = Time.now
-        timer.save
+    timer = Timer.find params[:id]
+    if can? :update, timer
+      if params[:act] == 'play'
+        if timer.started_at.nil?
+          timer.started_at = Time.now
+          timer.save
+        end
+      elsif params[:act] == 'pause'
+        if timer.started_at.present?
+          timer.seconds = timer.seconds.to_i + (Time.now - timer.started_at).to_i
+          timer.started_at = nil
+          timer.save
+        end
+      else
+        timer.update(timer_params)
       end
-    elsif params[:act] == 'pause'
-      if timer.started_at.present?
-        timer.seconds = timer.seconds.to_i + (Time.now - timer.started_at).to_i
-        timer.started_at = nil
-        timer.save
-      end
+      render json: { timer: timer_json_data(timer) }
     else
-      timer.update(timer_params)
+      render json: { error: 'access denied' }, status: :unprocessable_entity
     end
-    render json: { timer: timer_json_data(timer) }
   end
 
   def destroy
-    timer = current_user.timers.find params[:id]
-    return_data = timer_json_data timer, 'finished'
-    timer.destroy
+    timer = Timer.find params[:id]
+    if can? :destroy, timer
+      return_data = timer_json_data timer, 'finished'
+      timer.destroy
 
-    respond_to do |format|
-      format.html do
-        redirect_to timers_path
+      respond_to do |format|
+        format.html do
+          redirect_to timers_path
+        end
+        format.json do
+          render json: { timer: return_data }
+        end
       end
-      format.json do
-        render json: { timer: return_data }
+    else
+      respond_to do |format|
+        format.html do
+          flash[:alert] = "Access denied"
+          redirect_to timers_path
+        end
+        format.json do
+          render json: { error: 'access denied' }, status: :unprocessable_entity
+        end
       end
     end
   end
